feature:达人带货榜添加筛选

+++ models/es/dy_author_take_goods.go
@@ -8,10 +8,10 @@ type DyAuthorTakeGoods struct {
 	RoomCover        string  `json:"room_cover"`
 	CreateTime       int64   `json:"create_time"`
 	DiscoverTime     int64   `json:"discover_time"`
-	PredictSales     int     `json:"predict_sales"`
+	PredictSales     float64 `json:"predict_sales"`
 	PredictGmv       float64 `json:"predict_gmv"`
-	RealSales        int     `json:"real_sales"`
-	RealGmv          int     `json:"real_gmv"`
+	RealSales        float64 `json:"real_sales"`
+	RealGmv          float64 `json:"real_gmv"`
 	MaxUserCount     int     `json:"max_user_count"`
 	Nickname         string  `json:"nickname"`
 	ShortId          string  `json:"short_id"`
